latte-csbot-admin/upload_file/requirements.txt
# Ingestion Pipeline Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# PDF Processing
pymupdf>=1.24.0

# HTTP Client
httpx>=0.27.0

# Supabase
supabase>=2.0.0

# Document Processing
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0
Pillow>=10.0.0
odfpy>=1.4.0
fonttools>=4.0.0

# LlamaIndex
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0